fixup "os/vfs.h: read loop in Readonly_file::read" (-Wconversion on 32-bit)

+++ repos/os/include/os/vfs.h
@@ -492,7 +492,7 @@ class Genode::Readonly_file : public File
 				if (read_bytes == 0)
 					break;
 
-				total += read_bytes;
+				total += size_t(read_bytes);
 			}
 
 			return total;
